+++ src/world/Season.java
@@ -4,30 +4,30 @@ import ui.*;
 
 public class Season {
 
-	public static final double FREEZING_TEMPURATURE = 0.33;
-	public static final double MELTING_TEMPURATURE = 0.43;
+	public static final float FREEZING_TEMPURATURE = 0.33f;
+	public static final float MELTING_TEMPURATURE = 0.43f;
 	public static final int SEASON_DURATION = 6000;
-	public static double[] winter;
-	public static double[] summer;
-	public static double getSeason2() {
+	public static float[] winter;
+	public static float[] summer;
+	public static float getSeason2() {
 		int season =  (World.ticks + SEASON_DURATION*1/2)%(SEASON_DURATION*2);
-		return Math.abs(SEASON_DURATION - season) / (double)SEASON_DURATION;
+		return Math.abs(SEASON_DURATION - season) / (float)SEASON_DURATION;
 	}
 	public static double getSeason4() {
 		return (World.ticks + SEASON_DURATION*1/2)%(SEASON_DURATION*2) / (double)SEASON_DURATION;
 	}
 
-	private static final double SNOWY_POLES_RATIO = 0.04;
+	private static final float SNOWY_POLES_RATIO = 0.04f;
 	public static void makeSeasonArrays(int mapheight) {
-		summer = new double[mapheight];
-		winter = new double[mapheight];
+		summer = new float[mapheight];
+		winter = new float[mapheight];
 		int northPole = (int) (summer.length * SNOWY_POLES_RATIO);
 		int southPole = (int) (summer.length - summer.length * SNOWY_POLES_RATIO);
 		
 		int winterPoint = summer.length;
 		
 		for(int i = 0; i < summer.length; i++) {
-			double snowyPoles = 0;
+			float snowyPoles = 0;
 			if(i < northPole) {
 				snowyPoles += (1.0*northPole - i) / northPole;
 			}
@@ -35,7 +35,7 @@ public class Season {
 //				snowyPoles += (1.0*i - southPole) / northPole;
 //			}
 			summer[i] = snowyPoles;
-			double winterSeason = 0;
+			float winterSeason = 0;
 			if(i < winterPoint) {
 				winterSeason += (1.0 * winterPoint - i) / winterPoint;
 			}
